refactor: rename AttemptResult.php to AttemptSolutions.php and TestPassResult.php to AttemptResult.php to improve the separation between question results and the overall result of an attempt
