+++ components/ILIAS/Test/src/Results/Presentation/AttemptResultsTable.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 
 namespace ILIAS\Test\Results\Presentation;
 
-use ILIAS\Test\Results\Data\AttemptResult;
+use ILIAS\Test\Results\Data\AttemptSolutions;
 use ILIAS\Test\Results\Data\QuestionResult;
 use ILIAS\UI\Component\Table\Presentation as PresentationTable;
 use ILIAS\UI\Component\Table\PresentationRow;
@@ -48,16 +48,16 @@ class AttemptResultsTable
     protected PresentationTable $table;
 
     public function __construct(
-        UIFactory $ui_factory,
-        private UIRenderer $ui_renderer,
-        private Refinery $refinery,
-        private HTTPService $http,
-        DataFactory $data_factory,
-        Language $lng,
-        private AttemptResult $test_results,
-        Settings $settings,
-        string $title,
-        bool $for_print
+        UIFactory                $ui_factory,
+        private UIRenderer       $ui_renderer,
+        private Refinery         $refinery,
+        private HTTPService      $http,
+        DataFactory              $data_factory,
+        Language                 $lng,
+        private AttemptSolutions $test_results,
+        Settings                 $settings,
+        string                   $title,
+        bool                     $for_print
     ) {
         list($mode, $sortation) = $this->getViewControlsParameter();
         $results = $this->applyControls($mode, $sortation, $test_results->getQuestionResults());
