+++ components/ILIAS/Test/src/Results/Presentation/Factory.php
@@ -20,7 +20,7 @@ declare(strict_types=1);
 
 namespace ILIAS\Test\Results\Presentation;
 
-use ILIAS\Test\Results\Data\AttemptResult;
+use ILIAS\Test\Results\Data\AttemptSolutions;
 use ILIAS\Test\Results\Data\TestOverview;
 use ILIAS\UI\Factory as UIFactory;
 use ILIAS\UI\Renderer as UIRenderer;
@@ -44,10 +44,10 @@ class Factory
     }
 
     public function getAttemptResultsPresentationTable(
-        AttemptResult $attempt_result,
-        Settings $settings,
-        string $title,
-        bool $for_print
+        AttemptSolutions $attempt_result,
+        Settings         $settings,
+        string           $title,
+        bool             $for_print
     ): AttemptResultsTable {
         return new AttemptResultsTable(
             $this->ui_factory,
