+++ components/ILIAS/Test/tests/Results/Data/AttemptResultTest.php
@@ -18,7 +18,7 @@
 
 namespace ILIAS\Test\Tests\Results\Data;
 
-use ILIAS\Test\Results\Data\AttemptResult;
+use ILIAS\Test\Results\Data\AttemptSolutions;
 use ILIAS\Test\Results\Data\QuestionResult;
 
 class AttemptResultTest extends \ilTestBaseTestCase
@@ -33,7 +33,7 @@ class AttemptResultTest extends \ilTestBaseTestCase
     #[\PHPUnit\Framework\Attributes\DataProvider('getActiveIdDataProvider')]
     public function testGetActiveId(int $IO): void
     {
-        $ilTestPassResult = new AttemptResult(
+        $ilTestPassResult = new AttemptSolutions(
             $IO,
             0,
             []
@@ -53,7 +53,7 @@ class AttemptResultTest extends \ilTestBaseTestCase
     #[\PHPUnit\Framework\Attributes\DataProvider('getPassDataProvider')]
     public function testGetAttempt(int $IO): void
     {
-        $ilTestPassResult = new AttemptResult(
+        $ilTestPassResult = new AttemptSolutions(
             0,
             $IO,
             []
@@ -74,7 +74,7 @@ class AttemptResultTest extends \ilTestBaseTestCase
     public function testGetQuestionResults(\Closure $IO): void
     {
         $IO = $IO($this);
-        $ilTestPassResult = new AttemptResult(
+        $ilTestPassResult = new AttemptSolutions(
             0,
             0,
             $IO
